experiment-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
neo4j==5.15.0
pinecone-client==3.0.0
openai==1.12.0
anthropic==0.18.1
pdfplumber==0.10.3
pillow==10.2.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
